+++ Tests/Test-PublicIP.tests.ps1
@@ -1,0 +1,50 @@
+Describe "Test-PublicIP PS$PSVersion" {
+
+    BeforeAll {
+        # Skipping the import when the module is already loaded avoids a second copy being loaded
+        if (-not (Get-Module -Name 'Subnet')) {
+            Import-Module $PSScriptRoot\..\Subnet
+        }
+    }
+
+    $PrivateIPs = '10.0.0.0', '10.255.255.255', '172.16.0.0', '172.31.255.255', '192.168.0.0', '192.168.255.255'
+
+    Context 'Private IPs' {
+
+        It "Should return $false for <_>" -TestCases $PrivateIPs {
+            InModuleScope Subnet -Parameters @{ IP = $_ } {
+                Test-PublicIP -IP $IP | Should -Be $false
+            }
+        }
+    }
+
+    Context 'Private IPs with pipeline input' {
+
+        It "Should return $false for <_>" -TestCases $PrivateIPs {
+            InModuleScope Subnet -Parameters @{ IP = $_ } {
+                $IP | Test-PublicIP | Should -Be $false
+            }
+        }
+    }
+
+
+    $PublicIPs = '9.255.255.255', '11.0.0.0', '172.15.255.255', '172.32.0.0', '192.167.255.255', '192.169.0.0'
+
+    Context 'Public IPs' {
+
+        It "Should return $true for <_>" -TestCases $PublicIPs {
+            InModuleScope Subnet -Parameters @{ IP = $_ } {
+                Test-PublicIP -IP $IP | Should -Be $true
+            }
+        }
+    }
+
+    Context 'Public IPs with pipeline input' {
+
+        It "Should return $true for <_>" -TestCases $PublicIPs {
+            InModuleScope Subnet -Parameters @{ IP = $_ } {
+                $IP | Test-PublicIP | Should -Be $true
+            }
+        }
+    }
+}